Updated Platino core
to conform to most recent Arduino core

+++ source2/avr/cores/arduino/Arduino.h
@@ -235,7 +235,8 @@ uint16_t makeWord(byte h, byte l);
 unsigned long pulseIn(uint8_t pin, uint8_t state, unsigned long timeout = 1000000L);
 unsigned long pulseInLong(uint8_t pin, uint8_t state, unsigned long timeout = 1000000L);
 
-void tone(uint8_t _pin, unsigned int frequency, unsigned long duration = 0);
+// CPV: made frequency unsigned long instead of unsigned int to allow for higher frequencies.
+void tone(uint8_t _pin, unsigned long frequency, unsigned long duration = 0);
 void noTone(uint8_t _pin);
 
 // WMath prototypes
